+++ unmarshal.go
@@ -127,68 +127,6 @@ type ChunkResult struct {
 	err      error
 }
 
-func UnmarshalTableIntoFast(table arrow.Table, resultHolders any) (returnErr error) {
-	defer func() {
-		if panicContents := recover(); panicContents != nil {
-			detail := "details irretrievable"
-			switch typedContents := panicContents.(type) {
-			case *reflect.ValueError:
-				detail = typedContents.Error()
-			case string:
-				detail = typedContents
-			}
-			returnErr = errors.Newf("exception occurred while unmarshalling result: %s", detail)
-		}
-	}()
-
-	numRows, err := internal.Int64ToInt(table.NumRows())
-	if err != nil {
-		return errors.Newf("table too large to unmarshal, found %d rows", table.NumRows())
-	}
-
-	slicePtr := reflect.ValueOf(resultHolders)
-	if slicePtr.Kind() != reflect.Ptr {
-		return fmt.Errorf(
-			"result holder should be a pointer to a slice of structs, "+
-				"got '%s' instead",
-			slicePtr.Kind(),
-		)
-	}
-
-	slice := reflect.Indirect(slicePtr)
-	if slice.Kind() != reflect.Slice {
-		return fmt.Errorf(
-			"result holder should be a pointer to a slice of structs, "+
-				"got '%s' instead",
-			slice.Kind(),
-		)
-	}
-
-	sliceElemType := slice.Type().Elem()
-	if sliceElemType.Kind() != reflect.Struct {
-		return fmt.Errorf(
-			"result holder should be a pointer to a slice of structs, "+
-				"got a pointer to a slice of '%s' instead",
-			sliceElemType.Kind(),
-		)
-	}
-
-	allMemo := internal.AllNamespaceMemo
-	if err := internal.PopulateAllNamespaceMemo(sliceElemType); err != nil {
-		return errors.Wrap(err, "building namespace memo")
-	}
-
-	if slice.Len() != numRows {
-		slice.Set(reflect.MakeSlice(slice.Type(), numRows, numRows))
-	}
-
-	if err := internal.MapTableToStructs(table, &slice, allMemo); err != nil {
-		return errors.Wrap(err, "mapping table to structs")
-	}
-
-	return nil
-}
-
 func unmarshalTableInto(table arrow.Table, resultHolders any) (returnErr error) {
 	defer func() {
 		if panicContents := recover(); panicContents != nil {
@@ -243,7 +181,7 @@ func unmarshalTableInto(table arrow.Table, resultHolders any) (returnErr error) 
 		return nil
 	}
 
-	scope, err := buildScope(colls.Keys(rows[0]))
+	scope, err := internal.BuildScope(colls.Keys(rows[0]))
 	if err != nil {
 		return errors.Wrap(err, "building deserialization scope")
 	}
@@ -255,7 +193,7 @@ func unmarshalTableInto(table arrow.Table, resultHolders any) (returnErr error) 
 
 	structName := sliceElemType.Name()
 	namespace := internal.ChalkpySnakeCase(structName)
-	nsScope := scope.children[namespace]
+	nsScope := scope.Children[namespace]
 
 	var rowToStruct func(map[Fqn]any) (*reflect.Value, error)
 	if nsScope != nil {
@@ -284,7 +222,7 @@ func unmarshalTableInto(table arrow.Table, resultHolders any) (returnErr error) 
 		type namespaceMetaT struct {
 			fieldIdx  int
 			namespace string
-			scope     *scopeTrie
+			scope     *internal.InitScope
 			memo      *internal.NamespaceMemo
 		}
 
@@ -298,7 +236,7 @@ func unmarshalTableInto(table arrow.Table, resultHolders any) (returnErr error) 
 						"If attempting multi-namespace unmarshalling, please pass in a pointer to a struct whose fields are all "+
 						"structs (not struct pointers) corresponding to the output namespaces. The problematic field is '%s' of type '%s'.",
 					structName,
-					colls.Keys(scope.children),
+					colls.Keys(scope.Children),
 					fieldMeta.Name,
 					fieldMeta.Type.Name(),
 				)
@@ -306,7 +244,7 @@ func unmarshalTableInto(table arrow.Table, resultHolders any) (returnErr error) 
 
 			fieldNamespace := internal.ChalkpySnakeCase(fieldMeta.Type.Name())
 
-			fieldScope := scope.children[fieldNamespace]
+			fieldScope := scope.Children[fieldNamespace]
 			if fieldScope == nil {
 				return errors.Newf(
 					"Please make sure you're unmarshalling into the correct struct. Attempted single namespace "+
@@ -315,7 +253,7 @@ func unmarshalTableInto(table arrow.Table, resultHolders any) (returnErr error) 
 					structName,
 					fieldMeta.Name,
 					fieldMeta.Type.Name(),
-					colls.Keys(scope.children),
+					colls.Keys(scope.Children),
 				)
 			}
 
@@ -430,15 +368,7 @@ func unmarshalTableInto(table arrow.Table, resultHolders any) (returnErr error) 
 //
 //	}
 func UnmarshalTableInto(table arrow.Table, resultHolders any) error {
-	return UnmarshalTableIntoFast(table, resultHolders)
-}
-
-func buildScope(fqns []string) (*scopeTrie, error) {
-	root := &scopeTrie{}
-	for _, fqn := range fqns {
-		root.addStr(fqn)
-	}
-	return root, nil
+	return internal.UnmarshalTableIntoFast(table, resultHolders)
 }
 
 /*
@@ -473,7 +403,7 @@ func UnmarshalInto(resultHolder any, fqnToValue map[Fqn]any) (returnErr error) {
 	if err := internal.PopulateAllNamespaceMemo(reflect.ValueOf(resultHolder).Elem().Type()); err != nil {
 		return errors.Wrap(err, "building namespace memo")
 	}
-	scope, err := buildScope(colls.Keys(fqnToValue))
+	scope, err := internal.BuildScope(colls.Keys(fqnToValue))
 	if err != nil {
 		return errors.Wrap(err, "building scope for initializing result holder struct")
 	}
@@ -482,7 +412,7 @@ func UnmarshalInto(resultHolder any, fqnToValue map[Fqn]any) (returnErr error) {
 	structValue := holderValue.Elem()
 	structName := structValue.Type().Name()
 	namespace := internal.ChalkpySnakeCase(structName)
-	nsScope := scope.children[namespace]
+	nsScope := scope.Children[namespace]
 
 	if nsScope != nil {
 		// Single namespace unmarshalling
@@ -491,7 +421,7 @@ func UnmarshalInto(resultHolder any, fqnToValue map[Fqn]any) (returnErr error) {
 				"Attempted to unmarshal into the feature struct '%s', "+
 					"but results are from these feature class(es) '%v'",
 				structName,
-				colls.Keys(scope.children),
+				colls.Keys(scope.Children),
 			)
 		}
 
@@ -521,14 +451,14 @@ func UnmarshalInto(resultHolder any, fqnToValue map[Fqn]any) (returnErr error) {
 					"If attempting multi-namespace unmarshalling, please pass in a pointer to a struct whose fields are all "+
 					"structs (not struct pointers) corresponding to the output namespaces. The problematic field is '%s' of type '%s'.",
 				structName,
-				colls.Keys(scope.children),
+				colls.Keys(scope.Children),
 				fieldMeta.Name,
 				field.Type().Name(),
 			)
 		}
 		fieldNamespace := internal.ChalkpySnakeCase(field.Type().Name())
 
-		fieldNsScope := scope.children[fieldNamespace]
+		fieldNsScope := scope.Children[fieldNamespace]
 		if fieldNsScope == nil {
 			return errors.Newf(
 				"Please make sure you're unmarshalling into the correct struct. Attempted single namespace "+
@@ -537,7 +467,7 @@ func UnmarshalInto(resultHolder any, fqnToValue map[Fqn]any) (returnErr error) {
 				structName,
 				fieldMeta.Name,
 				field.Type().Name(),
-				colls.Keys(scope.children),
+				colls.Keys(scope.Children),
 			)
 		}
 
@@ -568,12 +498,12 @@ func thinUnmarshalInto(
 	structValue reflect.Value,
 	fqnToValue map[Fqn]any,
 	namespace string,
-	namespaceScope *scopeTrie,
+	namespaceScope *internal.InitScope,
 	namespaceMemo *internal.NamespaceMemo,
 	allMemo *internal.AllNamespaceMemoT,
 ) (returnErr error) {
 	remoteFeatureMap := map[string][]reflect.Value{}
-	if err := initRemoteFeatureMap(
+	if err := internal.InitRemoteFeatureMap(
 		remoteFeatureMap,
 		structValue,
 		namespace,
